requirements.txt
coverage==5.0.3
Django==3.0
djangorestframework==3.11
factory_boy==2.12.0
psycopg2==2.8.4